Place git blame line at end of line
When set to right_align it was blocking some of the actual code if the
window was too small to show the code and line blame text one after
another. This way it always goes the end of the line which means the
line blame text will get cut off rather than the code getting covered by
the line blame text and for me that's preferable.